Move full runtime traits into full:: module
