+++ crates/relayer-framework/src/base/one_for_all/traits/runtime.rs
@@ -1,6 +1,5 @@
 use async_trait::async_trait;
 use core::fmt::Debug;
-use core::future::Future;
 use core::time::Duration;
 
 use crate::base::core::traits::sync::Async;
@@ -20,56 +19,4 @@ pub trait OfaBaseRuntime: Clone + Async {
     fn now(&self) -> Self::Time;
 
     fn duration_since(time: &Self::Time, other: &Self::Time) -> Duration;
-}
-
-#[async_trait]
-pub trait OfaFullRuntime: OfaBaseRuntime {
-    type Sender<T>: Async
-    where
-        T: Async;
-
-    type Receiver<T>: Async
-    where
-        T: Async;
-
-    type SenderOnce<T>: Async
-    where
-        T: Async;
-
-    type ReceiverOnce<T>: Async
-    where
-        T: Async;
-
-    fn spawn<F>(&self, task: F)
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static;
-
-    fn new_channel<T>() -> (Self::Sender<T>, Self::Receiver<T>)
-    where
-        T: Async;
-
-    fn send<T>(sender: &Self::Sender<T>, value: T) -> Result<(), Self::Error>
-    where
-        T: Async;
-
-    async fn receive<T>(receiver: &Self::Receiver<T>) -> Result<T, Self::Error>
-    where
-        T: Async;
-
-    async fn try_receive<T>(receiver: &Self::Receiver<T>) -> Result<Option<T>, Self::Error>
-    where
-        T: Async;
-
-    fn new_channel_once<T>() -> (Self::SenderOnce<T>, Self::ReceiverOnce<T>)
-    where
-        T: Async;
-
-    fn send_once<T>(sender: Self::SenderOnce<T>, value: T) -> Result<(), Self::Error>
-    where
-        T: Async;
-
-    async fn receive_once<T>(receiver: Self::ReceiverOnce<T>) -> Result<T, Self::Error>
-    where
-        T: Async;
 }
