+++ crates/relayer-framework/src/base/runtime/traits/mod.rs
@@ -1,8 +1,5 @@
-pub mod channel;
-pub mod channel_once;
 pub mod log;
 pub mod mutex;
 pub mod runtime;
 pub mod sleep;
-pub mod spawn;
 pub mod time;
