+++ crates/relayer-framework/src/full/batch/worker.rs
@@ -6,17 +6,17 @@ use crate::base::chain::traits::types::HasIbcChainTypes;
 use crate::base::core::traits::sync::Async;
 use crate::base::relay::traits::target::ChainTarget;
 use crate::base::relay::traits::types::HasRelayTypes;
-use crate::base::runtime::traits::channel::{CanUseChannels, HasChannelTypes};
-use crate::base::runtime::traits::channel_once::{CanUseChannelsOnce, HasChannelOnceTypes};
 use crate::base::runtime::traits::log::{HasLogger, LevelDebug};
 use crate::base::runtime::traits::runtime::HasRuntime;
 use crate::base::runtime::traits::sleep::CanSleep;
-use crate::base::runtime::traits::spawn::{HasSpawner, Spawner};
 use crate::base::runtime::traits::time::HasTime;
 use crate::full::batch::traits::channel::HasMessageBatchReceiver;
 use crate::full::batch::traits::send_messages_from_batch::CanSendIbcMessagesFromBatchWorker;
 use crate::full::batch::types::aliases::{BatchSubmission, EventResultSender};
 use crate::full::batch::types::config::BatchConfig;
+use crate::full::runtime::traits::channel::{CanUseChannels, HasChannelTypes};
+use crate::full::runtime::traits::channel_once::{CanUseChannelsOnce, HasChannelOnceTypes};
+use crate::full::runtime::traits::spawn::{HasSpawner, Spawner};
 use crate::std_prelude::*;
 
 pub struct BatchMessageWorker<Relay, Target>
