+++ crates/relayer-framework/src/full/mod.rs
@@ -7,4 +7,5 @@ pub mod batch;
 pub mod filter;
 pub mod one_for_all;
 pub mod relay;
+pub mod runtime;
 pub mod telemetry;
