+++ crates/relayer-framework/src/full/one_for_all/impls/mod.rs
@@ -3,4 +3,5 @@ pub mod chain;
 pub mod filter;
 pub mod preset;
 pub mod relay;
+pub mod runtime;
 pub mod telemetry;
