+++ crates/relayer-framework/src/full/one_for_all/impls/runtime/channel.rs
@@ -1,10 +1,10 @@
 use async_trait::async_trait;
 
 use crate::base::core::traits::sync::Async;
-use crate::base::one_for_all::traits::runtime::OfaFullRuntime;
 use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
-use crate::base::runtime::traits::channel::{CanCreateChannels, CanUseChannels, HasChannelTypes};
-use crate::base::runtime::traits::channel_once::{
+use crate::full::one_for_all::traits::runtime::OfaFullRuntime;
+use crate::full::runtime::traits::channel::{CanCreateChannels, CanUseChannels, HasChannelTypes};
+use crate::full::runtime::traits::channel_once::{
     CanCreateChannelsOnce, CanUseChannelsOnce, HasChannelOnceTypes,
 };
 use crate::std_prelude::*;
