+++ crates/relayer-framework/src/full/one_for_all/impls/runtime/spawn.rs
@@ -1,8 +1,8 @@
 use core::future::Future;
 
-use crate::base::one_for_all::traits::runtime::OfaFullRuntime;
 use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
-use crate::base::runtime::traits::spawn::{HasSpawner, Spawner};
+use crate::full::one_for_all::traits::runtime::OfaFullRuntime;
+use crate::full::runtime::traits::spawn::{HasSpawner, Spawner};
 use crate::std_prelude::*;
 
 impl<Runtime: OfaFullRuntime> HasSpawner for OfaRuntimeWrapper<Runtime> {
