+++ crates/relayer-framework/src/full/one_for_all/traits/mod.rs
@@ -1,3 +1,4 @@
 pub mod chain;
 pub mod relay;
+pub mod runtime;
 pub mod telemetry;
