+++ crates/relayer-framework/src/full/one_for_all/traits/relay.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 
 use crate::base::one_for_all::traits::relay::OfaBaseRelay;
-use crate::base::one_for_all::traits::runtime::OfaFullRuntime;
+use crate::full::one_for_all::traits::runtime::OfaFullRuntime;
 use crate::full::one_for_all::types::runtime::aliases::{MessageBatchReceiver, MessageBatchSender};
 use crate::std_prelude::*;
