+++ crates/relayer-framework/src/full/one_for_all/traits/runtime.rs
@@ -1,0 +1,58 @@
+use async_trait::async_trait;
+use core::future::Future;
+
+use crate::base::core::traits::sync::Async;
+use crate::base::one_for_all::traits::runtime::OfaBaseRuntime;
+use crate::std_prelude::*;
+
+#[async_trait]
+pub trait OfaFullRuntime: OfaBaseRuntime {
+    type Sender<T>: Async
+    where
+        T: Async;
+
+    type Receiver<T>: Async
+    where
+        T: Async;
+
+    type SenderOnce<T>: Async
+    where
+        T: Async;
+
+    type ReceiverOnce<T>: Async
+    where
+        T: Async;
+
+    fn spawn<F>(&self, task: F)
+    where
+        F: Future + Send + 'static,
+        F::Output: Send + 'static;
+
+    fn new_channel<T>() -> (Self::Sender<T>, Self::Receiver<T>)
+    where
+        T: Async;
+
+    fn send<T>(sender: &Self::Sender<T>, value: T) -> Result<(), Self::Error>
+    where
+        T: Async;
+
+    async fn receive<T>(receiver: &Self::Receiver<T>) -> Result<T, Self::Error>
+    where
+        T: Async;
+
+    async fn try_receive<T>(receiver: &Self::Receiver<T>) -> Result<Option<T>, Self::Error>
+    where
+        T: Async;
+
+    fn new_channel_once<T>() -> (Self::SenderOnce<T>, Self::ReceiverOnce<T>)
+    where
+        T: Async;
+
+    fn send_once<T>(sender: Self::SenderOnce<T>, value: T) -> Result<(), Self::Error>
+    where
+        T: Async;
+
+    async fn receive_once<T>(receiver: Self::ReceiverOnce<T>) -> Result<T, Self::Error>
+    where
+        T: Async;
+}
